+++ CHANGELOG.md
@@ -4,6 +4,10 @@
  - update `jsoneditor.min.js` from `2.6.1` to `2.9.1`
  - better labeling and colors for `data_config` and `condition` buttons based on if a configuration is set or not
  - add new internal style `jobConfig`
+ - refactor user input data to be with types: `internal`, `external` instead of `dynamic`, `static`
+
+### Plugins
+ - `fitrockr` plugin should be updated to v1.0.4
 
 # v5.11.3
 ### Security Update